+++ service/flex/v1/plugin_releases/api_op_client.go
@@ -1,0 +1,17 @@
+// Package plugin_releases contains auto-generated files. DO NOT MODIFY
+package plugin_releases
+
+import "github.com/RJPearson94/twilio-sdk-go/client"
+
+// Client for managing plugin release resources
+// See https://www.twilio.com/docs/flex/developer/plugins/api/release for more details
+type Client struct {
+	client *client.Client
+}
+
+// New creates a new instance of the plugin releases client
+func New(client *client.Client) *Client {
+	return &Client{
+		client: client,
+	}
+}